fix: edge case
If the bound is smaller than the weight of some input literal that has
not been included in the objective encoding yet, the encoding will _not_
include said literal, but add it as an assumption directly. In this case
we cannot use the totalizer output as a reified objective bound
directly.